move quant method to dedicated file
